# webhook-event-handler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Webhook Event Handler API",
    "description": "\n## Webhook Integration API\n\nBidirectional webhook event processing service that synchronises **knowledge\narticles**, **case data**, and **usage analytics** between eGain and\nSalesforce Service Cloud.\n\n---\n\n### Integration Flows\n\n| # | Endpoint | Body `event` | Direction | Downstream |\n|---|----------|--------------|-----------|------------|\n| 1 | `POST /webhooks/article-published` | `article.published` | eGain → Salesforce | Salesforce `Knowledge__kav` sObject |\n| 2 | `POST /webhooks/case-closed` | `case.closed` | Salesforce → eGain | Content Gap API |\n| 3 | `POST /webhooks/article-viewed` | `article.viewed` | eGain → eGain Analytics | Analytics API / Kafka |\n\n---\n\n### Authentication\n\nEvery inbound webhook **must** include two headers:\n\n| Header | Format |\n|--------|--------|\n| `X-Signature` | `sha256=<64-char hex digest>` |\n| `X-Timestamp` | Unix epoch seconds (±5 min tolerance) |\n\n**Signing algorithm**\n```\nsigning_payload = f\"{X-Timestamp}.{raw_request_body}\"\nsignature       = HMAC-SHA256(shared_secret, signing_payload)\nheader_value    = \"sha256=\" + signature.hex()\n```\n\n---\n\n",
    "contact": {
      "name": "Nazneen Ansari",
      "email": "nazneen.ansari379@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
...
    }
  ],
  "paths": {
    "/webhooks/article-published": {
      "post": {
        "tags": [
          "Webhooks - Knowledge Articles"
        ],
        "summary": "Receive article-published event",
        "description": "Called by the **eGain CMS** whenever a knowledge article is published or updated.\nThe service transforms the payload and enqueues a push to **Salesforce Knowledge**\n(`Knowledge__kav` sObject).\n\nRequest body must include `\"event\": \"article.published\"`, `tenantId` (eGain tenant),\nand `appId` (the application triggering the webhook).\n\n**Processing flow:**\n1. eGain publishes article → fires this webhook\n2. HMAC-SHA256 signature is verified\n3. System validates the request body\n4. Event is enqueued → `202 Accepted` returned immediately\n5. Worker transforms the article and calls Salesforce Knowledge API",
        "operationId": "receive_article_published_webhooks_article_published_post",
        "parameters": [
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "sha256=<hex-digest> of `{timestamp}.{raw_body}`",
              "examples": [
                "sha256=4b1f3c8a9e2d7f0b6a3e5c8d1f4a7b2e9c6d0a3f7e1b4c8d2f5a9e0b3c6d1f4"
              ],
              "title": "X-Signature"
            },
            "description": "sha256=<hex-digest> of `{timestamp}.{raw_body}`"
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unix epoch seconds at time of signing.",
              "examples": [
                "1717243200"
              ],
              "title": "X-Timestamp"
            },
            "description": "Unix epoch seconds at time of signing."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticlePublishedWebhook"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Event accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden – tenant not authorised for this operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/case-closed": {
      "post": {
        "tags": [
          "Webhooks - Cases"
        ],
        "summary": "Receive case-closed event",
        "description": "Called by **Salesforce** (Outbound Message or Platform Event) when a support case\nis closed. The service analyses whether the case was resolved with existing\nknowledge articles, computes a **content-gap**, and records a gap entry\nin eGain.\n\nRequest body must include `\"event\": \"case.closed\"`, `tenantId` (eGain tenant),\nand `appId` (the application triggering the webhook).\n\n**Processing flow:**\n1. Salesforce closes case → fires this webhook\n2. HMAC-SHA256 signature verified\n3. Event enqueued → `202 Accepted` returned\n4. Worker takes the data and stores it in a database for scheduled batch processing for content-gap analysis.",
        "operationId": "receive_case_closed_webhooks_case_closed_post",
        "parameters": [
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "sha256=<hex-digest> of `{timestamp}.{raw_body}`",
              "examples": [
                "sha256=4b1f3c8a9e2d7f0b6a3e5c8d1f4a7b2e9c6d0a3f7e1b4c8d2f5a9e0b3c6d1f4"
              ],
              "title": "X-Signature"
            },
            "description": "sha256=<hex-digest> of `{timestamp}.{raw_body}`"
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unix epoch seconds at time of signing.",
              "examples": [
                "1717243200"
              ],
              "title": "X-Timestamp"
            },
            "description": "Unix epoch seconds at time of signing."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseClosedWebhook"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Event accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden – tenant not authorised for this operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/article-viewed": {
      "post": {
        "tags": [
          "Webhooks - Analytics"
        ],
        "summary": "Receive article-viewed analytics event",
        "description": "Called by the **eGain portal** on every article page view. Events are streamed\nto the **eGain Analytics API** (or Kafka/Kinesis in high-volume deployments).\n\nRequest body must include `\"event\": \"article.viewed\"`, `tenantId` (eGain tenant),\nand `appId` (the application triggering the webhook).\n\nThis is a **high-throughput** endpoint — expect bursts of hundreds of events\nper second during peak support hours.\n\n**Processing flow:**\n1. eGain fires view event in real time\n2. Signature verified, body validated\n3. Event enqueued → `202 Accepted` returned\n4. Worker batches and forwards to `POST /api/v1/analytics/events`\n5. Analytics pipeline updates article usage metrics in eGain",
        "operationId": "receive_article_viewed_webhooks_article_viewed_post",
        "parameters": [
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "sha256=<hex-digest> of `{timestamp}.{raw_body}`",
              "examples": [
                "sha256=4b1f3c8a9e2d7f0b6a3e5c8d1f4a7b2e9c6d0a3f7e1b4c8d2f5a9e0b3c6d1f4"
              ],
              "title": "X-Signature"
            },
            "description": "sha256=<hex-digest> of `{timestamp}.{raw_body}`"
          },
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Unix epoch seconds at time of signing.",
              "examples": [
                "1717243200"
              ],
              "title": "X-Timestamp"
            },
            "description": "Unix epoch seconds at time of signing."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleViewedWebhook"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Analytics event accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden – tenant not authorised for this operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "accepted",
            "examples": [
              "accepted"
            ]
          },
          "eventId": {
            "type": "string",
            "format": "uuid",
            "title": "Eventid",
            "description": "Assigned UUID for status polling.",
            "examples": [
              "3fa85f64-5717-4562-b3fc-2c963f66afa6"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Article publish event queued for processing."
            ]
          }
        },
        "type": "object",
        "required": [
          "eventId",
          "message"
        ],
        "title": "AcceptedResponse",
        "description": "Returned immediately (HTTP 202) for every inbound webhook."
      },
      "ArticleData": {
        "properties": {
          "articleId": {
            "type": "string",
            "title": "Articleid",
            "description": "Unique article identifier in eGain.",
            "examples": [
              "ARTICLE-1001"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "title": "Title",
            "description": "Article title — maps to Salesforce Knowledge `Title` field.",
            "examples": [
              "How to reset your password"
            ]
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Full HTML article body. Maps to Salesforce `ArticleBody`.",
            "examples": [
              "<p>Step 1: Click <b>Forgot Password</b>...</p>"
            ]
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "BCP-47 language tag. Maps to Salesforce `Language`.",
            "default": "en-US",
            "examples": [
              "en-US",
              "fr-FR",
              "de-DE"
            ]
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Article version string stored in custom field `eGain__Source_Article_Version__c`.",
            "examples": [
              "2.0"
            ]
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories",
            "description": "eGain category names. Joined with ';' into Salesforce `eGain__Categories__c`.",
            "examples": [
              [
                "Account Management",
                "Security"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "articleId",
          "title",
          "urlName",
          "content"
        ],
        "title": "ArticleData",
        "description": "Knowledge article payload from eGain CMS."
      },
      "ArticlePublishedWebhook": {
        "properties": {
          "webhookId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhookid",
            "description": "Unique delivery ID assigned by eGain. Use for idempotency checks.",
            "examples": [
              "wh_01HZK8M3N2PQRS"
            ]
          },
          "event": {
            "type": "string",
            "title": "Event",
            "description": "Must be exactly `article.published`.",
            "examples": [
              "article.published"
            ]
          },
          "tenantId": {
            "type": "string",
            "title": "Tenantid",
            "description": "eGain tenant identifier. Required for all webhook calls.",
            "examples": [
              "acme-corp"
            ]
          },
          "appId": {
            "type": "string",
            "title": "Appid",
            "description": "Identifier of the application triggering this webhook.",
            "examples": [
              "egain"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "ISO-8601 event time in UTC.",
            "examples": [
              "2024-06-01T12:00:00Z"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ArticleData"
          }
        },
        "type": "object",
        "required": [
          "event",
          "tenantId",
          "appId",
          "timestamp",
          "data"
        ],
        "title": "ArticlePublishedWebhook",
        "description": "Webhook envelope sent by eGain when a knowledge article is published or updated.\nTriggers a push to **Salesforce Knowledge** (`Knowledge__kav` sObject)."
      },
...
      "ArticleViewedWebhook": {
        "properties": {
          "webhookId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhookid",
            "examples": [
              "wh_03HZK0N5O4RSTU"
            ]
          },
          "event": {
            "type": "string",
            "title": "Event",
            "description": "Must be exactly `article.viewed`.",
            "examples": [
              "article.viewed"
            ]
          },
          "tenantId": {
            "type": "string",
            "title": "Tenantid",
            "description": "eGain tenant identifier. Required for all webhook calls.",
            "examples": [
              "acme-corp"
            ]
          },
          "appId": {
            "type": "string",
            "title": "Appid",
            "description": "Identifier of the application triggering this webhook.",
            "examples": [
              "egain-portal-v2"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "examples": [
              "2024-06-01T17:00:00Z"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ArticleViewData"
          }
        },
        "type": "object",
        "required": [
          "event",
          "tenantId",
          "appId",
          "timestamp",
          "data"
        ],
        "title": "ArticleViewedWebhook",
        "description": "Webhook envelope sent by eGain on every article page view.\nHigh-throughput endpoint — expect bursts of hundreds of events per second.\nEvents are streamed to the eGain Analytics API (or Kafka/Kinesis in production)."
      },
      "CaseClosedWebhook": {
        "properties": {
          "event": {
            "type": "string",
            "title": "Event",
            "description": "Must be exactly `case.closed`.",
            "examples": [
              "case.closed"
            ]
          },
          "tenantId": {
            "type": "string",
            "title": "Tenantid",
            "description": "eGain tenant identifier. Required for all webhook calls.",
            "examples": [
              "acme-corp"
            ]
          },
          "appId": {
            "type": "string",
            "title": "Appid",
            "description": "Identifier of the application triggering this webhook.",
            "examples": [
              "salesforce-service-cloud"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "ISO-8601 event time in UTC.",
            "examples": [
              "2024-06-01T15:00:00Z"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/CaseData"
          }
        },
        "type": "object",
        "required": [
          "event",
          "tenantId",
          "appId",
          "timestamp",
          "data"
        ],
        "title": "CaseClosedWebhook",
        "description": "Webhook envelope from Salesforce (Outbound Message or Platform Event)\nwhen a support case is closed. Triggers content-gap analysis in eGain."
      },
      "CaseData": {
        "properties": {
          "caseId": {
            "type": "string",
            "title": "Caseid",
            "description": "Salesforce Case record ID (15 or 18-char Salesforce ID).",
            "examples": [
              "5003000000ABC123AAA"
            ]
          },
          "caseNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casenumber",
            "description": "Human-readable case number shown in Salesforce UI.",
            "examples": [
              "00001234"
            ]
          },
          "subject": {
            "type": "string",
            "maxLength": 255,
            "title": "Subject",
            "description": "Case subject line.",
            "examples": [
              "Cannot log into portal"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Full case description / customer problem statement.",
            "examples": [
              "User says password reset email not arriving."
            ]
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution",
            "description": "Agent resolution notes. Null = no resolution recorded.",
            "examples": [
              "Sent user the password reset knowledge article."
            ]
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Salesforce case category / record type.",
            "default": "Uncategorised",
            "examples": [
              "Account Management"
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/Priority",
            "description": "Salesforce case priority.",
            "default": "Medium",
            "examples": [
              "Medium"
            ]
          },
          "articleIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Articleids",
            "description": "eGain article IDs **suggested** to the agent during the case.",
            "examples": [
              [
                "ARTICLE-1001",
                "ARTICLE-1005"
              ]
            ]
          },
          "viewedArticleIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Viewedarticleids",
            "description": "eGain article IDs the agent actually **opened/read**.",
            "examples": [
              [
                "ARTICLE-1001"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "caseId",
          "subject",
          "description"
        ],
        "title": "CaseData",
        "description": "Salesforce case payload delivered on case closure."
      },
...
      "UnauthorizedError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Machine-readable error code. Always `UNAUTHORIZED` for 401 responses.",
            "examples": [
              "UNAUTHORIZED"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable description of the authentication failure.",
            "examples": [
              "Missing or invalid HMAC signature.",
              "X-Timestamp header is absent or outside the ±5 min tolerance.",
              "X-Signature header is missing."
            ]
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "UnauthorizedError",
        "description": "Returned on HTTP 401 when authentication credentials are missing or the HMAC signature cannot be verified."
      },
      "ForbiddenError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Machine-readable error code. Always `FORBIDDEN` for 403 responses.",
            "examples": [
              "FORBIDDEN"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable description of why the request was denied.",
            "examples": [
              "Tenant acme-corp is not authorised to publish articles.",
              "appId egain-cms-v3 does not have permission to call this endpoint.",
              "Insufficient permissions to access the Dead Letter Queue."
            ]
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ForbiddenError",
        "description": "Returned on HTTP 403 when the caller is authenticated but not permitted to perform the requested operation."
      },
      "InternalServerError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Machine-readable error code. Always `INTERNAL_SERVER_ERROR` for 500 responses.",
            "examples": [
              "INTERNAL_SERVER_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable description of the server-side failure.",
            "examples": [
              "An unexpected error occurred. Please try again later.",
              "Queue service is temporarily unavailable.",
              "Failed to persist event to the database."
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional diagnostic detail. Omitted in production to avoid leaking internal state.",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "InternalServerError",
        "description": "Returned on HTTP 500 when an unexpected server-side error prevents the request from completing."
      },
...
      "Priority": {
        "type": "string",
        "enum": [
          "Low",
          "Medium",
          "High",
          "Critical"
        ],
        "title": "Priority"
      },
...
    }
  },
...
}